Update 1.0.2
Fix artifact float value error

+++ enkapy/model/artifact.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel, Field
 
 class ArtifactProperty(BaseModel):
     mainPropId: str = Field('', alias='mainPropId')
-    statValue: int = Field(0, alias='statValue')
+    statValue: float = Field(0, alias='statValue')
     appendPropId: str = Field('', alias='appendPropId')
 
     @property
